Structured Programming Language
In this file, you will get all the documents for SPL courses. I additionally added all the practice problems for the SPL course, those will help you to practice in a proper way and of course, if you do practice all the problems set you will have a good grade and best of luck.